+++ core/src/sync/mod.rs
@@ -7,6 +7,7 @@ mod synchronization_graph;
 mod synchronization_protocol_handler;
 mod synchronization_service;
 mod synchronization_state;
+mod request_manager;
 
 pub use self::{
     error::{Error, ErrorKind},
@@ -23,7 +24,7 @@ pub use self::{
         SharedSynchronizationService, SynchronizationService,
     },
     synchronization_state::{
-        SentTransactionContainer, SynchronizationPeerState,
+        SynchronizationPeerState,
         SynchronizationState,
     },
 };
@@ -31,4 +32,47 @@ pub use self::{
 pub mod random {
     use rand;
     pub fn new() -> rand::ThreadRng { rand::thread_rng() }
+}
+
+pub mod msg_sender {
+    use network::{NetworkContext, PeerId, throttling::THROTTLING_SERVICE};
+    use message::Message;
+    use priority_send_queue::SendQueuePriority;
+    use cfx_bytes::Bytes;
+    use network::Error as NetworkError;
+
+    pub fn send_message(
+        io: &NetworkContext, peer: PeerId, msg: &Message,
+        priority: SendQueuePriority,
+    ) -> Result<(), NetworkError>
+    {
+        send_message_with_throttling(io, peer, msg, priority, false)
+    }
+
+    pub fn send_message_with_throttling(
+        io: &NetworkContext, peer: PeerId, msg: &Message,
+        priority: SendQueuePriority, throttling_disabled: bool,
+    ) -> Result<(), NetworkError>
+    {
+        if !throttling_disabled && msg.is_size_sensitive() {
+            if let Err(e) = THROTTLING_SERVICE.read().check_throttling() {
+                debug!("Throttling failure: {:?}", e);
+                return Err(e);
+            }
+        }
+        let mut raw = Bytes::new();
+        raw.push(msg.msg_id().into());
+        raw.extend(msg.rlp_bytes().iter());
+        if let Err(e) = io.send(peer, raw, priority) {
+            debug!("Error sending message: {:?}", e);
+            return Err(e);
+        };
+        debug!(
+            "Send message({}) to {:?}",
+            msg.msg_id(),
+            io.get_peer_node_id(peer)
+        );
+        Ok(())
+    }
+
 }
